+++ Tests/Utils/Arrays/README IsShortArrayTests.md
@@ -1,0 +1,2 @@
+The Arrays::isShortArray() method is tested together with the Lists::isShortList() method.
+The tests for this can be found in the Lists folder.
